(PC-33385)[API] fix: fixed circular import, moved BannerMetaModel and VenueTypeCode from native serialization module to core offerers schemas
