+++ api/tests/routes/pro/patch_draft_offer_test.py
@@ -7,6 +7,7 @@ from pcapi.connectors import api_adresse
 from pcapi.core.categories import subcategories_v2 as subcategories
 from pcapi.core.geography import models as geography_models
 import pcapi.core.offerers.factories as offerers_factories
+from pcapi.core.offerers.schemas import VenueTypeCode
 import pcapi.core.offers.factories as offers_factories
 from pcapi.core.offers.models import Offer
 from pcapi.core.offers.models import OfferStatus
@@ -14,7 +15,6 @@ import pcapi.core.providers.factories as providers_factories
 from pcapi.core.providers.repository import get_provider_by_local_class
 from pcapi.core.testing import override_features
 import pcapi.core.users.factories as users_factories
-from pcapi.routes.native.v1.serialization.offerers import VenueTypeCode
 from pcapi.utils.date import format_into_utc_date
